feat: set up do-not-land and if-change-then-change (#5)
do-not-land asserts that if you have a 'DO NOT LAND' somewhere in the code then it can't get checked in; with hold-the-line this works out to a 'DO NOT LAND' in the diff

if-change-then-change keeps different code blocks in sync. has to go through libgit2 and eventually rely on `${upstream}` from trunk, because it needs to know the exact set of modified lines, not just the set of modified files. for some frustrating reason `diff_tree_to_workdir_with_index` doesn't recognize untracked files, even though i explicitly set it to true; apparently in trunk we go thru a `git status` API instead. (i'm not married to the current syntax, but it's the best that we have right now; it also suffers from a number of structural deficiencies, e.g. doesn't support multiple other files nor block labels)

there are still a number of antipatterns in here (e.g. the unwrap in main.rs) but for now i think i'm reasonably happy with how the impl looks. tests are certainly terrible and need to be written, but that's a separate problem.