feat: fix release process and rename horton -> trunk-toolbox (#11)
Per discussion with David and Chris, we're going with a more generic name because we don't expect this tool to build its own brand identity independent of trunk.

Also allow controlling the released version using `HORTON_RELEASE` (`TOOLBOX_RELEASE` reads weird without context). This has to be done manually in build.rs; the Rust ecosystem normally assumes that this is done via `package.version` in Cargo.toml, but that's a weird way to do things.

Sneak in a bugfix for resolving --upstream; we also now handle arbitrary hex OIDs.