fix(ictc): write tests, fix moved blocks & LFS, and more (#16)
Teach if-change-then-change to respect moved if-change-then-change
blocks and LFS.

* Previously, we asked libgit2 for the hunks and blindly forwarded them
to ICTC as the new&modified line ranges in a file. This is problematic
because if lines are added before an ICTC block, libgit2 will say that
the ICTC block moved, because its line numbers changed, even though its
contents didn't change. As an example, let's say the contents of
`foo.file` change from this

    ```
    1   alpha
    2   beta gamma
    3   // IfChange
    4   block contents
    5   // ThenChange other.file
    ```

    to this

    ```
    1   alpha
    2   beta
    3   gamma
    4   // IfChange
    5   block contents
    6   // ThenChange other.file
    ```

The `DiffHunk` that libgit2 returns will span old lines 2 through 5 and
new lines 2 through 6, even though the contents of the ICTC block
haven't changed.

To handle this case - which is honestly a pretty basic case - we instead
construct our own hunks using the line-by-line diff that libgit2 hands
back to us, which exclude lines whose contents have not changed between
the upstream and current, even if their line numbers have changed.

* We also add handling for LFS files. Today, toolbox errors out when it
is presented with a diff that includes an LFS file, because it tries to
parse the LFS file as a UTF-8 text file.

This doesn't happen when just handling normal binary files, because when
we use libgit2 to enumerate the set of files to lint, we rely on libgit2
to exclude binary files. Git itself, however, does not recognize LFS
files as binary files - it thinks of them as text files whose attributes
specify that Git should access using the LFS driver.

To fix this, we just teach toolbox to check the attributes of every file
in the diff, and to exclude files with a filter=lfs git attribute. (We
could also use diff=lfs; there's no particular reason to choose one over
another.)

We also make the following changes as well:

  * Add test coverage for if-change-then-change
* Start using fluent test assertions via `spectral`, so that error
messages in failing tests are actually informative about what's failing
* Stop using content_inspector to filter out binary files, since the new
LFS handling logic applies at file set resolution time.
* Support arbitrary revspecs (e.g. `HEAD^`, which is not a valid
reference, or hash prefixes) for `--upstream`, by going through
`revparse_single`