Add new TODO rule (#40)
Adds a rule (default disabled) that checks for TODOs and FIXMEs. This is
very similar to DO_NOT_LAND but outputs at a lower severity and default
config disabled.

The intention is for this to be our OOTB solution for a TODO linter (as
compared with per-language linter plugins and abusing linting
spellcheckers). I merged most of the logic with DNL so that we wouldn't
have to read through files twice.