Add "no-curly-quotes" rule to the codebase.  (#56)
Adds support for nocurlyquotes that will find instances of curlyquotes
in a file and replace with correct non-curly form. We had instances of
this in our code base and it caused some issues with marketing site.

Fun little rule. Interestingly - this rule exposes a bug in how we apply
on `trunk check` sarif fixes for utf-8 non-ascii characters. For now
this rule will stay disabled until we have a fix.

Fix bug where multiple fixes in a single issue were being turned into
multiple issues (which is not what the spec or code was intending to
do).